stupid

+++ go/cmd/gcsobjtable/main.go
@@ -177,7 +177,7 @@ func (s *GCSObjServer) GetObjectLocations(ctx context.Context, req *pb.ObjectLoc
 	locations := make(map[uint64]*pb.LocationByteTuple)
 	for u := range req.Args {
 		if _,ok := s.objectLocations[uint64(u)]; ok {
-			locations[uint64(u)] = &pb.LocationByteTuple{Location: s.objectLocations[u]}
+			locations[uint64(u)] = &pb.LocationByteTuple{Location: s.objectLocations[uint64(u)]}
 		}
 		
 	}
